src/requirements.txt
absl-py==0.7.0
altgraph==0.16.1
astor==0.7.1
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.1
coverage==4.5.2
cycler==0.10.0
Django==2.1.5
future==0.17.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
isort==4.3.4
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
macholib==1.11
Markdown==3.0.1
matplotlib==3.0.2
mccabe==0.6.1
more-itertools==5.0.0
nose==1.3.7
numpy==1.16.0
opencv-python==4.0.0.21
paho-mqtt==1.4.0
pandas==0.24.0
pefile==2018.8.8
Pillow==5.4.1
pluggy==0.8.1
protobuf==3.6.1
psycopg2==2.7.7
py==1.7.0
PyInstaller==3.4
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.14
pyqt5-tools==5.11.3.1.4
pytest==4.1.1
pytest-cov==2.6.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
pywin32-ctypes==0.2.0
PyYAML==3.13
requests==2.21.0
scipy==1.2.0
sip==4.19.8
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.4.1
Werkzeug==0.14.1
wrapt==1.11.1
